requirements.txt
torch==2.2.2
torchvision==0.17.2
diffusers==0.30.2
transformers==4.45.2
accelerate==0.33.0
safetensors==0.4.3
huggingface-hub==0.34.3
pillow==11.0.0
tqdm==4.67.1
python-dotenv==1.0.1
flask
redis
rq
boto3
python-dotenv
pillow
torch
diffusers
transformers
accelerate
safetensors
